+++ server/src/model/contents.rs
@@ -5,11 +5,11 @@ use super::users::SimpleUser;
 
 #[derive(FromRow, Serialize, Deserialize)]
 pub struct Content {
-    pub r#type: i64,
+    pub r#type: i32,
     pub title: String,
     pub source: String,
-    pub author: i64,
-    pub topic: i64,
+    pub author: i32,
+    pub topic: i32,
     pub description: String,
     pub cover_image: String,
     pub up_date: sqlx::types::chrono::NaiveDate,
